refactor: move funboxes to a shared package (@miodec) (#6063)
