+++ backend/package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "@date-fns/utc": "1.2.0",
     "@monkeytype/contracts": "workspace:*",
+    "@monkeytype/funbox": "workspace:*",
     "@monkeytype/util": "workspace:*",
     "@ts-rest/core": "3.51.0",
     "@ts-rest/express": "3.51.0",
