+++ frontend/__tests__/root/config.spec.ts
@@ -332,10 +332,8 @@ describe("Config", () => {
     expect(Config.setFavThemes([stringOfLength(51)])).toBe(false);
   });
   it("setFunbox", () => {
-    expect(Config.setFunbox("one")).toBe(true);
-    expect(Config.setFunbox("one#two")).toBe(true);
-    expect(Config.setFunbox("one#two#")).toBe(true);
-    expect(Config.setFunbox(stringOfLength(100))).toBe(true);
+    expect(Config.setFunbox("mirror")).toBe(true);
+    expect(Config.setFunbox("mirror#58008")).toBe(true);
 
     expect(Config.setFunbox(stringOfLength(101))).toBe(false);
   });
