+++ frontend/package.json
@@ -70,6 +70,7 @@
   "dependencies": {
     "@date-fns/utc": "1.2.0",
     "@monkeytype/contracts": "workspace:*",
+    "@monkeytype/funbox": "workspace:*",
     "@monkeytype/util": "workspace:*",
     "@ts-rest/core": "3.51.0",
     "canvas-confetti": "1.5.1",
