+++ frontend/src/ts/ready.ts
@@ -1,14 +1,13 @@
-import Config from "./config";
 import * as Misc from "./utils/misc";
 import * as MonkeyPower from "./elements/monkey-power";
 import * as MerchBanner from "./elements/merch-banner";
 import * as CookiesModal from "./modals/cookies";
 import * as ConnectionState from "./states/connection";
 import * as AccountButton from "./elements/account-button";
-import * as FunboxList from "./test/funbox/funbox-list";
 //@ts-expect-error
 import Konami from "konami";
 import * as ServerConfiguration from "./ape/server-configuration";
+import { getActiveFunboxes } from "./test/funbox/list";
 
 $((): void => {
   Misc.loadCSS("/css/slimselect.min.css", true);
@@ -21,9 +20,9 @@ $((): void => {
   $("body").css("transition", "background .25s, transform .05s");
   MerchBanner.showIfNotClosedBefore();
   setTimeout(() => {
-    FunboxList.get(Config.funbox).forEach((it) =>
-      it.functions?.applyGlobalCSS?.()
-    );
+    for (const fb of getActiveFunboxes()) {
+      fb.functions?.applyGlobalCSS?.();
+    }
   }, 500); //this approach will probably bite me in the ass at some point
 
   $("#app")
