+++ frontend/src/ts/test/funbox/memory-funbox-timer.ts
@@ -1,5 +1,3 @@
-import * as TestWords from "../test-words";
-
 let memoryTimer: number | null = null;
 let memoryInterval: NodeJS.Timeout | null = null;
 
@@ -30,9 +28,9 @@ export function reset(): void {
   hide();
 }
 
-export function start(): void {
+export function start(time: number): void {
   reset();
-  memoryTimer = Math.round(Math.pow(TestWords.words.length, 1.2));
+  memoryTimer = time;
   update(memoryTimer);
   show();
   memoryInterval = setInterval(() => {
