+++ frontend/src/ts/test/words-generator.ts
@@ -1,5 +1,4 @@
 import Config, * as UpdateConfig from "../config";
-import * as FunboxList from "./funbox/funbox-list";
 import * as CustomText from "./custom-text";
 import * as Wordset from "./wordset";
 import QuotesController, {
@@ -17,6 +16,7 @@ import * as Arrays from "../utils/arrays";
 import * as TestState from "../test/test-state";
 import * as GetText from "../utils/generate";
 import { FunboxWordOrder, LanguageObject } from "../utils/json-data";
+import { getActiveFunboxes } from "./funbox/list";
 
 function shouldCapitalize(lastChar: string): boolean {
   return /[?!.؟]/.test(lastChar);
@@ -35,9 +35,8 @@ export async function punctuateWord(
 
   const lastChar = Strings.getLastChar(previousWord);
 
-  const funbox = FunboxList.get(Config.funbox).find(
-    (f) => f.functions?.punctuateWord
-  );
+  const funbox = getActiveFunboxes()?.find((fb) => fb.functions?.punctuateWord);
+
   if (funbox?.functions?.punctuateWord) {
     return funbox.functions.punctuateWord(word);
   }
@@ -302,25 +301,25 @@ async function applyEnglishPunctuationToWord(word: string): Promise<string> {
 }
 
 function getFunboxWordsFrequency(): Wordset.FunboxWordsFrequency | undefined {
-  const wordFunbox = FunboxList.get(Config.funbox).find(
-    (f) => f.functions?.getWordsFrequencyMode
+  const funbox = getActiveFunboxes().find(
+    (fb) => fb.functions?.getWordsFrequencyMode
   );
-  if (wordFunbox?.functions?.getWordsFrequencyMode) {
-    return wordFunbox.functions.getWordsFrequencyMode();
+  if (funbox?.functions?.getWordsFrequencyMode) {
+    return funbox.functions.getWordsFrequencyMode();
   }
   return undefined;
 }
 
 async function getFunboxSection(): Promise<string[]> {
   const ret = [];
-  const sectionFunbox = FunboxList.get(Config.funbox).find(
-    (f) => f.functions?.pullSection
-  );
-  if (sectionFunbox?.functions?.pullSection) {
-    const section = await sectionFunbox.functions.pullSection(Config.language);
+
+  const funbox = getActiveFunboxes().find((fb) => fb.functions?.pullSection);
+
+  if (funbox?.functions?.pullSection) {
+    const section = await funbox.functions.pullSection(Config.language);
 
     if (section === false || section === undefined) {
-      UpdateConfig.toggleFunbox(sectionFunbox.name);
+      UpdateConfig.toggleFunbox(funbox.name);
       throw new Error("Failed to pull section");
     }
 
@@ -339,19 +338,18 @@ function getFunboxWord(
   wordIndex: number,
   wordset?: Wordset.Wordset
 ): string {
-  const wordFunbox = FunboxList.get(Config.funbox).find(
-    (f) => f.functions?.getWord
-  );
-  if (wordFunbox?.functions?.getWord) {
-    word = wordFunbox.functions.getWord(wordset, wordIndex);
+  const funbox = getActiveFunboxes()?.find((fb) => fb.functions?.getWord);
+
+  if (funbox?.functions?.getWord) {
+    word = funbox.functions.getWord(wordset, wordIndex);
   }
   return word;
 }
 
 function applyFunboxesToWord(word: string): string {
-  for (const f of FunboxList.get(Config.funbox)) {
-    if (f.functions?.alterText) {
-      word = f.functions.alterText(word);
+  for (const fb of getActiveFunboxes()) {
+    if (fb.functions?.alterText) {
+      word = fb.functions.alterText(word);
     }
   }
   return word;
@@ -384,7 +382,7 @@ function applyLazyModeToWord(word: string, language: LanguageObject): string {
 
 export function getWordOrder(): FunboxWordOrder {
   const wordOrder =
-    FunboxList.get(Config.funbox)
+    getActiveFunboxes()
       .find((f) => f.properties?.find((fp) => fp.startsWith("wordOrder")))
       ?.properties?.find((fp) => fp.startsWith("wordOrder")) ?? "";
 
@@ -409,7 +407,7 @@ export function getWordsLimit(): number {
   }
 
   const funboxToPush =
-    FunboxList.get(Config.funbox)
+    getActiveFunboxes()
       .find((f) => f.properties?.find((fp) => fp.startsWith("toPush")))
       ?.properties?.find((fp) => fp.startsWith("toPush:")) ?? "";
 
@@ -607,8 +605,8 @@ export async function generateWords(
     hasNewline: false,
   };
 
-  const sectionFunbox = FunboxList.get(Config.funbox).find(
-    (f) => f.functions?.pullSection
+  const sectionFunbox = getActiveFunboxes().find(
+    (fb) => fb.functions?.pullSection
   );
   isCurrentlyUsingFunboxSection =
     sectionFunbox?.functions?.pullSection !== undefined;
@@ -632,11 +630,9 @@ export async function generateWords(
     wordList = wordList.reverse();
   }
 
-  const wordFunbox = FunboxList.get(Config.funbox).find(
-    (f) => f.functions?.withWords
-  );
-  if (wordFunbox?.functions?.withWords) {
-    currentWordset = await wordFunbox.functions.withWords(wordList);
+  const funbox = getActiveFunboxes().find((fb) => fb.functions?.withWords);
+  if (funbox?.functions?.withWords) {
+    currentWordset = await funbox.functions.withWords(wordList);
   } else {
     currentWordset = await Wordset.withWords(wordList);
   }
