--- frontend/static/funbox/_list.json
@@ -1,204 +1,0 @@
-[
-  {
-    "name": "nausea",
-    "info": "I think I'm gonna be sick.",
-    "canGetPb": true
-  },
-  {
-    "name": "round_round_baby",
-    "info": "...right round, like a record baby. Right, round round round.",
-    "canGetPb": true
-  },
-  {
-    "name": "simon_says",
-    "info": "Type what simon says.",
-    "canGetPb": true
-  },
-  {
-    "name": "mirror",
-    "info": "Everything is mirrored!",
-    "canGetPb": true
-  },
-  {"name": "upside_down",
-  "info": "Everything is upside down!",
-  "canGetPb": true
-  },
-  {
-    "name": "tts",
-    "info": "Listen closely.",
-    "canGetPb": true
-  },
-  {
-    "name": "choo_choo",
-    "info": "All the letters are spinning!",
-    "canGetPb": true
-  },
-  {
-    "name": "arrows",
-    "info": "Play it on a pad!",
-    "canGetPb": false
-  },
-  {
-    "name": "rAnDoMcAsE",
-    "info": "I kInDa LiKe HoW iNeFfIcIeNt QwErTy Is.",
-    "canGetPb": false
-  },
-  {
-    "name": "capitals",
-    "info": "Capitalize Every Word.",
-    "canGetPb": false
-  },
-  {
-    "name": "layoutfluid",
-    "info": "Switch between layouts specified below proportionately to the length of the test.",
-    "canGetPb": true
-  },
-  {
-    "name": "earthquake",
-    "info": "Everybody get down! The words are shaking!",
-    "canGetPb": true
-  },
-  {
-    "name": "space_balls",
-    "info": "In a galaxy far far away.",
-    "canGetPb": true
-  },
-  {
-    "name": "gibberish",
-    "info": "Anvbuefl dizzs eoos alsb?",
-    "canGetPb": false
-  },
-  {
-    "name": "58008",
-    "alias": "numbers",
-    "info": "A special mode for accountants.",
-    "canGetPb": false
-  },
-  {
-    "name": "ascii",
-    "info": "Where was the ampersand again?. Only ASCII characters.",
-    "canGetPb": false
-  },
-  {
-    "name": "specials",
-    "info": "!@#$%^&*. Only special characters.",
-    "canGetPb": false
-  },
-  {
-    "name": "plus_zero",
-    "info": "React quickly! Only the current word is visible.",
-    "canGetPb": true
-  },
-  {
-    "name": "plus_one",
-    "info": "Only one future word is visible.",
-    "canGetPb": true
-  },
-  {
-    "name": "plus_two",
-    "info": "Only two future words are visible.",
-    "canGetPb": true
-  },
-  {
-    "name": "plus_three",
-    "info": "Only three future words are visible.",
-    "canGetPb": true
-  },
-  {
-    "name": "read_ahead_easy",
-    "info": "Only the current word is invisible.",
-    "canGetPb": true
-  },
-  {
-    "name": "read_ahead",
-    "info": "Current and the next word are invisible!",
-    "canGetPb": true
-  },
-  {
-    "name": "read_ahead_hard",
-    "info": "Current and the next two words are invisible!",
-    "canGetPb": true
-  },
-  {
-    "name": "memory",
-    "info": "Test your memory. Remember the words and type them blind.",
-    "canGetPb": true
-  },
-  {
-    "name": "nospace",
-    "info": "Whoneedsspacesanyway?",
-    "canGetPb": false
-  },
-  {
-    "name": "poetry",
-    "info": "Practice typing some beautiful prose.",
-    "canGetPb": false
-  },
-  {
-    "name": "wikipedia",
-    "info": "Practice typing wikipedia sections.",
-    "canGetPb": false
-  },
-  {
-    "name": "weakspot",
-    "info": "Focus on slow and mistyped letters.",
-    "canGetPb": false
-  },
-  {
-    "name": "pseudolang",
-    "info": "Nonsense words that look like the current language.",
-    "canGetPb": false
-  },
-  {
-    "name": "IPv4",
-    "alias": "network",
-    "info": "For sysadmins.",
-    "canGetPb": false
-  },
-  {
-    "name": "IPv6",
-    "alias": "network",
-    "info": "For sysadmins with a long beard.",
-    "canGetPb": false
-  },
-  {
-    "name": "binary",
-    "info": "01000010 01100101 01100101 01110000 00100000 01100010 01101111 01101111 01110000 00101110",
-    "canGetPb": false
-  },
-  {
-    "name": "hexadecimal",
-    "info": "0x38 0x20 0x74 0x69 0x6D 0x65 0x73 0x20 0x6D 0x6F 0x72 0x65 0x20 0x62 0x6F 0x6F 0x70 0x21",
-    "canGetPb": false
-  },
-  {
-    "name": "zipf",
-    "info": "Words are generated according to Zipf's law. (not all languages will produce Zipfy results, use with caution)",
-    "canGetPb": false
-  },
-  {
-    "name": "morse",
-    "info": "-.../././.--./ -.../---/---/.--./-.-.--/ ",
-    "canGetPb": false
-  },
-  {
-    "name": "crt",
-    "info": "Go back to the 1980s",
-    "canGetPb": true
-  },
-  {
-    "name": "backwards",
-    "info": "...sdrawkcab epyt ot yrt woN",
-    "canGetPb": true
-  },
-  {
-    "name": "ddoouubblleedd",
-    "info": "TTyyppee eevveerryytthhiinngg ttwwiiccee..",
-    "canGetPb": true
-  },
-  {
-    "name": "instant_messaging",
-    "info": "Who needs shift anyway?",
-    "canGetPb": false
-  }
-]
