+++ frontend/vitest.config.js
@@ -10,5 +10,12 @@ export default defineConfig({
     coverage: {
       include: ["**/*.ts"],
     },
+    deps: {
+      optimizer: {
+        web: {
+          include: ["@monkeytype/funbox"],
+        },
+      },
+    },
   },
 });
