+++ packages/funbox/src/list.ts
@@ -1,14 +1,8 @@
-export type FunboxMetadata = {
-  name: string;
-  canGetPb: boolean;
-  difficultyLevel: number;
-  properties?: string[];
-  frontendForcedConfig?: Record<string, string[] | boolean[]>;
-  frontendFunctions?: string[];
-};
+import { FunboxMetadata, FunboxName } from "./types";
 
-const FunboxList: FunboxMetadata[] = [
-  {
+const list: Record<FunboxName, FunboxMetadata> = {
+  "58008": {
+    description: "A special mode for accountants.",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "ignoresLayout", "noLetters"],
@@ -22,64 +16,69 @@ const FunboxList: FunboxMetadata[] = [
       "handleChar",
     ],
     name: "58008",
+    alias: "numbers",
   },
-  {
-    canGetPb: true,
-    difficultyLevel: 2,
-    frontendFunctions: ["applyCSS"],
-    name: "nausea",
-  },
-  {
-    canGetPb: true,
-    difficultyLevel: 3,
-    frontendFunctions: ["applyCSS"],
-    name: "round_round_baby",
-  },
-  {
-    canGetPb: true,
-    difficultyLevel: 1,
-    properties: ["changesWordsVisibility", "usesLayout"],
-    frontendForcedConfig: {
-      highlightMode: ["letter", "off"],
-    },
-    frontendFunctions: ["applyCSS", "applyConfig", "rememberSettings"],
-    name: "simon_says",
-  },
-  {
-    canGetPb: true,
-    difficultyLevel: 3,
-    frontendFunctions: ["applyCSS"],
+  mirror: {
     name: "mirror",
-  },
-  {
+    description: "Everything is mirrored!",
+    properties: ["hasCssFile"],
     canGetPb: true,
     difficultyLevel: 3,
-    frontendFunctions: ["applyCSS"],
-    name: "upside_down",
   },
-  {
+  upside_down: {
+    name: "upside_down",
+    description: "Everything is upside down!",
+    properties: ["hasCssFile"],
+    canGetPb: true,
+    difficultyLevel: 3,
+  },
+  nausea: {
+    name: "nausea",
+    description: "I think I'm gonna be sick.",
+    canGetPb: true,
+    difficultyLevel: 2,
+    properties: ["hasCssFile"],
+  },
+  round_round_baby: {
+    name: "round_round_baby",
+    description:
+      "...right round, like a record baby. Right, round round round.",
+    canGetPb: true,
+    difficultyLevel: 3,
+    properties: ["hasCssFile"],
+  },
+  simon_says: {
+    name: "simon_says",
+    description: "Type what simon says.",
     canGetPb: true,
     difficultyLevel: 1,
-    properties: ["changesWordsVisibility", "speaks"],
+    properties: ["hasCssFile", "changesWordsVisibility", "usesLayout"],
     frontendForcedConfig: {
       highlightMode: ["letter", "off"],
     },
-    frontendFunctions: [
-      "applyCSS",
-      "applyConfig",
-      "rememberSettings",
-      "toggleScript",
-    ],
-    name: "tts",
+    frontendFunctions: ["applyConfig", "rememberSettings"],
   },
-  {
+
+  tts: {
+    canGetPb: true,
+    difficultyLevel: 1,
+    properties: ["hasCssFile", "changesWordsVisibility", "speaks"],
+    frontendForcedConfig: {
+      highlightMode: ["letter", "off"],
+    },
+    frontendFunctions: ["applyConfig", "rememberSettings", "toggleScript"],
+    name: "tts",
+    description: "Listen closely.",
+  },
+  choo_choo: {
     canGetPb: true,
     difficultyLevel: 2,
-    properties: ["noLigatures", "conflictsWithSymmetricChars"],
-    frontendFunctions: ["applyCSS"],
+    properties: ["hasCssFile", "noLigatures", "conflictsWithSymmetricChars"],
     name: "choo_choo",
+    description: "All the letters are spinning!",
   },
-  {
+  arrows: {
+    description: "Play it on a pad!",
     canGetPb: false,
     difficultyLevel: 1,
     properties: [
@@ -96,7 +95,6 @@ const FunboxList: FunboxMetadata[] = [
     },
     frontendFunctions: [
       "getWord",
-      "applyConfig",
       "rememberSettings",
       "handleChar",
       "isCharCorrect",
@@ -105,21 +103,25 @@ const FunboxList: FunboxMetadata[] = [
     ],
     name: "arrows",
   },
-  {
+  rAnDoMcAsE: {
+    description: "I kInDa LiKe HoW iNeFfIcIeNt QwErTy Is.",
     canGetPb: false,
     difficultyLevel: 2,
     properties: ["changesCapitalisation"],
     frontendFunctions: ["alterText"],
     name: "rAnDoMcAsE",
   },
-  {
+  capitals: {
+    description: "Capitalize Every Word.",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["changesCapitalisation"],
     frontendFunctions: ["alterText"],
     name: "capitals",
   },
-  {
+  layoutfluid: {
+    description:
+      "Switch between layouts specified below proportionately to the length of the test.",
     canGetPb: true,
     difficultyLevel: 1,
     properties: ["changesLayout", "noInfiniteDuration"],
@@ -132,27 +134,30 @@ const FunboxList: FunboxMetadata[] = [
     ],
     name: "layoutfluid",
   },
-  {
+  earthquake: {
+    description: "Everybody get down! The words are shaking!",
     canGetPb: true,
     difficultyLevel: 1,
-    properties: ["noLigatures"],
-    frontendFunctions: ["applyCSS"],
+    properties: ["hasCssFile", "noLigatures"],
     name: "earthquake",
   },
-  {
+  space_balls: {
+    description: "In a galaxy far far away.",
     canGetPb: true,
     difficultyLevel: 0,
-    frontendFunctions: ["applyCSS"],
+    properties: ["hasCssFile"],
     name: "space_balls",
   },
-  {
+  gibberish: {
+    description: "Anvbuefl dizzs eoos alsb?",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "unspeakable"],
     frontendFunctions: ["getWord"],
     name: "gibberish",
   },
-  {
+  ascii: {
+    description: "Where was the ampersand again?. Only ASCII characters.",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "noLetters", "unspeakable"],
@@ -163,7 +168,8 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["getWord"],
     name: "ascii",
   },
-  {
+  specials: {
+    description: "!@#$%^&*. Only special characters.",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "noLetters", "unspeakable"],
@@ -174,61 +180,69 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["getWord"],
     name: "specials",
   },
-  {
+  plus_one: {
+    description: "Only one future word is visible.",
     canGetPb: true,
     difficultyLevel: 0,
     properties: ["changesWordsVisibility", "toPush:2", "noInfiniteDuration"],
     name: "plus_one",
   },
-  {
+  plus_zero: {
+    description: "React quickly! Only the current word is visible.",
     canGetPb: true,
     difficultyLevel: 1,
     properties: ["changesWordsVisibility", "toPush:1", "noInfiniteDuration"],
     name: "plus_zero",
   },
-  {
+  plus_two: {
+    description: "Only two future words are visible.",
     canGetPb: true,
     difficultyLevel: 0,
     properties: ["changesWordsVisibility", "toPush:3", "noInfiniteDuration"],
     name: "plus_two",
   },
-  {
+  plus_three: {
+    description: "Only three future words are visible.",
     canGetPb: true,
     difficultyLevel: 0,
     properties: ["changesWordsVisibility", "toPush:4", "noInfiniteDuration"],
     name: "plus_three",
   },
-  {
+  read_ahead_easy: {
+    description: "Only the current word is invisible.",
     canGetPb: true,
     difficultyLevel: 1,
-    properties: ["changesWordsVisibility"],
+    properties: ["changesWordsVisibility", "hasCssFile"],
     frontendForcedConfig: {
       highlightMode: ["letter", "off"],
     },
-    frontendFunctions: ["applyCSS", "rememberSettings", "handleKeydown"],
+    frontendFunctions: ["rememberSettings", "handleKeydown"],
     name: "read_ahead_easy",
   },
-  {
+  read_ahead: {
+    description: "Current and the next word are invisible!",
     canGetPb: true,
     difficultyLevel: 2,
-    properties: ["changesWordsVisibility"],
+    properties: ["changesWordsVisibility", "hasCssFile"],
     frontendForcedConfig: {
       highlightMode: ["letter", "off"],
     },
-    frontendFunctions: ["applyCSS", "rememberSettings", "handleKeydown"],
+    frontendFunctions: ["rememberSettings", "handleKeydown"],
     name: "read_ahead",
   },
-  {
+  read_ahead_hard: {
+    description: "Current and the next two words are invisible!",
     canGetPb: true,
     difficultyLevel: 3,
-    properties: ["changesWordsVisibility"],
+    properties: ["changesWordsVisibility", "hasCssFile"],
     frontendForcedConfig: {
       highlightMode: ["letter", "off"],
     },
-    frontendFunctions: ["applyCSS", "rememberSettings", "handleKeydown"],
+    frontendFunctions: ["rememberSettings", "handleKeydown"],
     name: "read_ahead_hard",
   },
-  {
+  memory: {
+    description: "Test your memory. Remember the words and type them blind.",
     canGetPb: true,
     difficultyLevel: 3,
     properties: ["changesWordsVisibility", "noInfiniteDuration"],
@@ -238,17 +252,19 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["applyConfig", "rememberSettings", "start", "restart"],
     name: "memory",
   },
-  {
+  nospace: {
+    description: "Whoneedsspacesanyway?",
     canGetPb: false,
     difficultyLevel: 0,
     properties: ["nospace"],
     frontendForcedConfig: {
       highlightMode: ["letter", "off"],
     },
-    frontendFunctions: ["applyConfig", "rememberSettings"],
+    frontendFunctions: ["rememberSettings"],
     name: "nospace",
   },
-  {
+  poetry: {
+    description: "Practice typing some beautiful prose.",
     canGetPb: false,
     difficultyLevel: 0,
     properties: ["noInfiniteDuration", "ignoresLanguage"],
@@ -259,7 +275,8 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["pullSection"],
     name: "poetry",
   },
-  {
+  wikipedia: {
+    description: "Practice typing wikipedia sections.",
     canGetPb: false,
     difficultyLevel: 0,
     properties: ["noInfiniteDuration", "ignoresLanguage"],
@@ -270,21 +287,25 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["pullSection"],
     name: "wikipedia",
   },
-  {
+  weakspot: {
+    description: "Focus on slow and mistyped letters.",
     canGetPb: false,
     difficultyLevel: 0,
     properties: ["changesWordsFrequency"],
     frontendFunctions: ["getWord"],
     name: "weakspot",
   },
-  {
+  pseudolang: {
+    description: "Nonsense words that look like the current language.",
     canGetPb: false,
     difficultyLevel: 0,
     properties: ["unspeakable", "ignoresLanguage"],
     frontendFunctions: ["withWords"],
     name: "pseudolang",
   },
-  {
+  IPv4: {
+    alias: "network",
+    description: "For sysadmins.",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "ignoresLayout", "noLetters"],
@@ -294,7 +315,9 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["getWord", "punctuateWord", "rememberSettings"],
     name: "IPv4",
   },
-  {
+  IPv6: {
+    alias: "network",
+    description: "For sysadmins with a long beard.",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "ignoresLayout", "noLetters"],
@@ -304,7 +327,9 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["getWord", "punctuateWord", "rememberSettings"],
     name: "IPv6",
   },
-  {
+  binary: {
+    description:
+      "01000010 01100101 01100101 01110000 00100000 01100010 01101111 01101111 01110000 00101110",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "ignoresLayout", "noLetters"],
@@ -315,7 +340,9 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["getWord"],
     name: "binary",
   },
-  {
+  hexadecimal: {
+    description:
+      "0x38 0x20 0x74 0x69 0x6D 0x65 0x73 0x20 0x6D 0x6F 0x72 0x65 0x20 0x62 0x6F 0x6F 0x70 0x21",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["ignoresLanguage", "ignoresLayout", "noLetters"],
@@ -325,51 +352,99 @@ const FunboxList: FunboxMetadata[] = [
     frontendFunctions: ["getWord", "punctuateWord", "rememberSettings"],
     name: "hexadecimal",
   },
-  {
+  zipf: {
+    description:
+      "Words are generated according to Zipf's law. (not all languages will produce Zipfy results, use with caution)",
     canGetPb: false,
     difficultyLevel: 0,
     properties: ["changesWordsFrequency"],
     frontendFunctions: ["getWordsFrequencyMode"],
     name: "zipf",
   },
-  {
+  morse: {
+    description: "-.../././.--./ -.../---/---/.--./-.-.--/ ",
     canGetPb: false,
     difficultyLevel: 1,
-    properties: ["ignoresLanguage", "ignoresLayout", "noLetters", "noSpace"],
+    properties: ["ignoresLanguage", "ignoresLayout", "noLetters", "nospace"],
     frontendFunctions: ["alterText"],
     name: "morse",
   },
-  {
+  crt: {
+    description: "Go back to the 1980s",
     canGetPb: true,
     difficultyLevel: 0,
-    properties: ["noLigatures"],
+    properties: ["hasCssFile", "noLigatures"],
+    frontendFunctions: ["applyGlobalCSS", "clearGlobal"],
     name: "crt",
   },
-  {
+  backwards: {
+    description: "...sdrawkcab epyt ot yrt woN",
     name: "backwards",
     properties: [
+      "hasCssFile",
       "noLigatures",
       "conflictsWithSymmetricChars",
       "wordOrder:reverse",
     ],
-    frontendFunctions: ["applyCSS"],
     canGetPb: true,
+    frontendFunctions: ["alterText"],
     difficultyLevel: 3,
   },
-  {
+  ddoouubblleedd: {
+    description: "TTyyppee eevveerryytthhiinngg ttwwiiccee..",
     canGetPb: true,
     difficultyLevel: 1,
     properties: ["noLigatures"],
     frontendFunctions: ["alterText"],
     name: "ddoouubblleedd",
   },
-  {
+  instant_messaging: {
+    description: "Who needs shift anyway?",
     canGetPb: false,
     difficultyLevel: 1,
     properties: ["changesCapitalisation"],
     frontendFunctions: ["alterText"],
     name: "instant_messaging",
   },
-];
+};
 
-export default FunboxList;
+export function getFunbox(name: FunboxName): FunboxMetadata;
+export function getFunbox(names: FunboxName[]): FunboxMetadata[];
+export function getFunbox(
+  nameOrNames: FunboxName | FunboxName[]
+): FunboxMetadata | FunboxMetadata[] {
+  if (Array.isArray(nameOrNames)) {
+    const out = nameOrNames.map((name) => getObject()[name]);
+
+    //@ts-expect-error
+    if (out.includes(undefined)) {
+      throw new Error("One of the funboxes is invalid: " + nameOrNames);
+    }
+
+    return out;
+  } else {
+    const out = getObject()[nameOrNames];
+
+    if (out === undefined) {
+      throw new Error("Invalid funbox name: " + nameOrNames);
+    }
+
+    return out;
+  }
+}
+
+export function getObject(): Record<FunboxName, FunboxMetadata> {
+  return list;
+}
+
+export function getList(): FunboxMetadata[] {
+  const out: FunboxMetadata[] = [];
+  for (const name of getFunboxNames()) {
+    out.push(list[name]);
+  }
+  return out;
+}
+
+function getFunboxNames(): FunboxName[] {
+  return Object.keys(list) as FunboxName[];
+}
