+++ packages/funbox/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "@monkeytype/typescript-config/base.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src",
+    "declaration": true,
+    "declarationMap": true,
+    "moduleResolution": "Bundler",
+    "module": "ES6",
+    "target": "ES2015",
+    "lib": ["es2016"]
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
